% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eribayes-package.R
\docType{package}
\name{eribayes-package}
\alias{eribayes}
\alias{eribayes-package}
\title{eribayes: Tools to simplify Bayesian model fitting}
\description{
|- The package *eribayes* supports Bayesian model fitting and workflow by providing wrappers for common needs that are compatible with multiple MCMC frontends for the Stan probabilistic programming language.
}
\author{
\strong{Maintainer}: Tom Shafer \email{tom.shafer@elderresearch.com}

Other contributors:
\itemize{
  \item Elder Research \email{support@elderresearch.com} [copyright holder]
}

}
\keyword{internal}
